\name{conversion.example}
\alias{conversion.example}
\docType{data}
\title{
Data file for data conversion example
}
\description{
Data file for data conversion example
}
\usage{data(conversion.example)}
\format{
  A data frame with 10 observations on the following 4 variables.
  \describe{
    \item{\code{Age}}{a numeric vector}
    \item{\code{State}}{a factor with levels \code{NJ} \code{NY} \code{TX} \code{VA}}
    \item{\code{Height}}{a numeric vector}
    \item{\code{Income}}{a numeric vector}
  }
}
\examples{
data(conversion.example)

}
\keyword{datasets}
