\name{factor.example}
\alias{factor.example}
\docType{data}
\title{
Example file for illustrating categorical variables
}
\description{
Example file for illustrating categorical variables
}
\usage{data(factor.example)}
\format{
  A data frame with 6 observations on the following 4 variables.
  \describe{
    \item{\code{Name}}{a factor with levels \code{Chang} \code{Esther} \code{Joe} \code{Paul} \code{Ram} \code{Yonah}}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Salary_level}}{a numeric vector}
    \item{\code{Rating}}{a numeric vector}
  }
}

\examples{
data(factor.example)
}

