\name{grocery}
\alias{grocery}
\docType{data}
\title{
Grocery
}
\description{
Test data set for affinity analysis
}
\usage{data(grocery)}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{item}}{a factor with levels \code{Moisturizer} \code{Paper} \code{Pen} \code{Soda} \code{Toothbrush} \code{Toothpaste}}
  }
}
\examples{
data(grocery)

}
\keyword{datasets}
