\name{hodar}
\docType{package}
\title{Supporting code for "{Data analytics with R: A hands-on approach}"}
\description{Supporting code for "{Data analytics with R: A hands-on approach}"}
\details{
\tabular{ll}{
Package: \tab hodar\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-08\cr
License: \tab GPL-2\cr
}
}
\author{
Viswa Viswanathan

Maintainer: Who to complain to <kv.viswana@gmail.com>

}

