\name{hodar.rpart.prune}
\alias{hodar.rpart.prune}

\title{
hodar.rpart.prune
}
\description{
Prunes the tree produced by rpart
}
\usage{
hodar.rpart.prune(model)
}

\arguments{
  \item{model}{
A model generated by rpart
}
}

\value{
Returns a pruned tree
}

\author{
Viswa Viswanathan
}

\examples{
library(hodar)
data(lcars)
# The following builds the complete tree
cmodel = hodar.rpart.build(Car_type ~ ., lcars, 2, 0)
# Plot it
hodar.rpart.plot(cmodel)
cmodel
# The following prunes the tree
# Warning: The call to the following function can produce different results on different 
# invocations because of random number generator effects.
pruned = hodar.rpart.prune(cmodel)
hodar.rpart.plot(pruned)
## The function is currently defined as
function (model) 
{
    cptab = model$cptable
    xval_errs = cptab[, "xerror"]
    ind = which.min(xval_errs)
    err_cutoff = cptab[ind, "xerror"] + cptab[ind, "xstd"]
    chosen_cp = max(cptab[cptab[, "xerror"] <= err_cutoff, "CP"])
    mod = prune(model, chosen_cp)
    mod
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
