\name{hodar.rpart.reg.plot}
\alias{hodar.rpart.reg.plot}

\title{
hodar.rpart.reg
}
\description{
Plots a regression tree generated by rpart
}
\usage{
hodar.rpart.reg.plot(model)
}

\arguments{
  \item{model}{
The tree model we want to plot
}
}

\value{
No return value -- just plots the tree.
}

\author{
Viswa Viswanathan}

\examples{
library(hodar)
data(education)
fit = rpart(expense ~ region + urban + income + under18, data = education, method = "anova")
hodar.rpart.reg.plot(fit)

## The function is currently defined as
function (model) 
{
    library(rpart.plot)
    prp(model, type = 2, nn = TRUE, fallen.leaves = TRUE, faclen = 4, 
        varlen = 8, shadow.col = "gray")
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
