\name{housing}
\alias{housing}
\docType{data}
\title{
Boston Housing data for illustrating various data analytics techniques
}
\description{
Boston Housing data for illustrating various data analytics techniques
}
\usage{data(housing)}
\format{
  A data frame with 506 observations on the following 14 variables.
  \describe{
    \item{\code{CRIM}}{a numeric vector}
    \item{\code{ZN}}{a numeric vector}
    \item{\code{INDUS}}{a numeric vector}
    \item{\code{CHAS}}{a numeric vector}
    \item{\code{NOX}}{a numeric vector}
    \item{\code{RM}}{a numeric vector}
    \item{\code{AGE}}{a numeric vector}
    \item{\code{DIS}}{a numeric vector}
    \item{\code{RAD}}{a numeric vector}
    \item{\code{TAX}}{a numeric vector}
    \item{\code{PTRATIO}}{a numeric vector}
    \item{\code{B}}{a numeric vector}
    \item{\code{LSTAT}}{a numeric vector}
    \item{\code{MEDV}}{a numeric vector}
  }
}

\source{
K. Bache and M. Lichman. UCI MAchine Learning Repository, retrieved, AUgust 2013. http://archive.ics.uci.edu/ml

D. Harrison and D. L. Rubinfeld. Hedonic process and the demand for clean air. J. Environ. Econ. and Management, pp 81-102, 1978.

}
\references{
D. Harrison and D. L. Rubinfeld. Hedonic process and the demand for clean air. J. Environ. Econ. and Management, pp 81-102, 1978.
}
\examples{
data(housing)
}
\keyword{datasets}