\name{hw}
\alias{hw}
\docType{data}
\title{
Example data.
}
\description{
Example data.}
\usage{data(hw)}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{\code{age}}{a numeric vector}
    \item{\code{height}}{a numeric vector}
  }
}

\examples{
data(hw)
## maybe str(hw) ; plot(hw) ...
}
\keyword{datasets}
