\name{lcars}
\alias{lcars}
\docType{data}
\title{
Data to illustrate various classification techniques.
}
\description{
Data to illustrate various classification techniques.
}
\usage{data(lcars)}
\format{
  A data frame with 24 observations on the following 3 variables.
  \describe{
    \item{\code{Income}}{a numeric vector}
    \item{\code{Family_size}}{a numeric vector}
    \item{\code{Car_type}}{a factor with levels \code{Luxury} \code{Non-Luxury}}
  }
}

\examples{
data(lcars)
## maybe str(lcars) ; plot(lcars) ...
}
\keyword{datasets}
