\name{vacation}
\alias{vacation}
\docType{data}
\title{
Data to illustrate regression and classification techniques.
}
\description{
Data to illustrate regression and classification techniques.
}
\usage{data(vacation)}
\format{
  A data frame with 40 observations on the following 3 variables.
  \describe{
    \item{\code{Income}}{a numeric vector}
    \item{\code{Family_size}}{a numeric vector}
    \item{\code{Result}}{a factor with levels \code{Buyer} \code{Non-buyer}}
  }
}

\examples{
data(vacation)
## maybe str(vacation) ; plot(vacation) ...
}
\keyword{datasets}
